Split out cost-related code into `input/commodity/cost.rs`
